added step 5, updated step 4, validated objectid

+++ routes/service/mongooseMethods.js
@@ -5,7 +5,9 @@ const getContacts = async () => {
 };
 
 const getContactByID = async (contactID) => {
-  return await ContactSchema.findOne({ _id: contactID });
+  if (contactID.match(/^[0-9a-fA-F]{24}$/)) {
+    return await ContactSchema.findOne({ _id: contactID });
+  }
 };
 
 const createContact = async (userData) => {
@@ -13,13 +15,25 @@ const createContact = async (userData) => {
 };
 
 const deleteContact = async (contactID) => {
-  return await ContactSchema.findByIdAndRemove({ _id: contactID });
+  if (contactID.match(/^[0-9a-fA-F]{24}$/)) {
+    return await ContactSchema.findByIdAndRemove({ _id: contactID });
+  }
 };
 
 const updateContact = async (contactID, userData) => {
-  return await ContactSchema.findByIdAndUpdate({ _id: contactID }, userData, {
-    new: true,
-  });
+  if (contactID.match(/^[0-9a-fA-F]{24}$/)) {
+    return await ContactSchema.findByIdAndUpdate({ _id: contactID }, userData, {
+      new: true,
+    });
+  }
+};
+
+const updateStatusContact = async (contactID, body) => {
+  if (contactID.match(/^[0-9a-fA-F]{24}$/)) {
+    return await ContactSchema.findByIdAndUpdate({ _id: contactID }, body, {
+      new: true,
+    });
+  }
 };
 
 module.exports = {
@@ -28,4 +42,5 @@ module.exports = {
   createContact,
   deleteContact,
   updateContact,
+  updateStatusContact,
 };
